Create BinaryImages.m
This Matlab script converts RGB .jpg files into black and white binary images and is used to pre-process large batches of images of rocks for shape/size characterization.  The script applies a simple thresholding technique based on contrast within the image and works best with images of rocks placed on a solid color background.